# CloudyFront/apispecs/swagger.yaml
openapi: 3.0.1
info:
  title: Kjkardum.CloudyBack.Api
  version: '1.0'
paths:
...
  /api/resource/BaseResource:
    get:
      tags:
        - BaseResource
      parameters:
        - name: Page
          in: query
          schema:
            type: integer
            format: int32
        - name: PageSize
          in: query
          schema:
            type: integer
            format: int32
        - name: FilterBy
          in: query
          schema:
            type: string
        - name: OrderBy
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/BaseResourceDtoPaginatedResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResourceDtoPaginatedResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/BaseResourceDtoPaginatedResponse'
...
  '/api/resource/BaseResource/{resourceId}/container':
    get:
      tags:
        - BaseResource
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ContainerDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ContainerDto'
  '/api/resource/BaseResource/{resourceId}/audit-log':
    get:
      tags:
        - BaseResource
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Page
          in: query
          schema:
            type: integer
            format: int32
        - name: PageSize
          in: query
          schema:
            type: integer
            format: int32
        - name: FilterBy
          in: query
          schema:
            type: string
        - name: OrderBy
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AuditLogEntryPaginatedResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogEntryPaginatedResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AuditLogEntryPaginatedResponse'
  '/api/resource/BaseResource/{resourceId}/prometheus':
    post:
      tags:
        - BaseResource
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryPrometheusQuery'
          text/json:
            schema:
              $ref: '#/components/schemas/QueryPrometheusQuery'
          application/*+json:
            schema:
              $ref: '#/components/schemas/QueryPrometheusQuery'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PrometheusResultDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PrometheusResultDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PrometheusResultDto'
  '/api/resource/BaseResource/{resourceId}/loki':
    post:
      tags:
        - BaseResource
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryLokiQuery'
          text/json:
            schema:
              $ref: '#/components/schemas/QueryLokiQuery'
          application/*+json:
            schema:
              $ref: '#/components/schemas/QueryLokiQuery'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PrometheusResultDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PrometheusResultDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PrometheusResultDto'
...
  /api/resource/WebApplicationResource:
    post:
      tags:
        - WebApplicationResource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebApplicationResourceCommand'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateWebApplicationResourceCommand'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateWebApplicationResourceCommand'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WebApplicationResourceDto'
            application/json:
              schema:
                $ref: '#/components/schemas/WebApplicationResourceDto'
            text/json:
              schema:
                $ref: '#/components/schemas/WebApplicationResourceDto'
  '/api/resource/WebApplicationResource/{id}':
    get:
      tags:
        - WebApplicationResource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WebApplicationResourceDto'
            application/json:
              schema:
                $ref: '#/components/schemas/WebApplicationResourceDto'
            text/json:
              schema:
                $ref: '#/components/schemas/WebApplicationResourceDto'
    delete:
      tags:
        - WebApplicationResource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  '/api/resource/WebApplicationResource/{id}/deploymentConfiguration':
    put:
      tags:
        - WebApplicationResource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebApplicationDeploymentConfigurationCommand'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateWebApplicationDeploymentConfigurationCommand'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateWebApplicationDeploymentConfigurationCommand'
      responses:
        '200':
          description: OK
  '/api/resource/WebApplicationResource/{id}/configuration':
    post:
      tags:
        - WebApplicationResource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyWebApplicationConfigItemCommand'
          text/json:
            schema:
              $ref: '#/components/schemas/ModifyWebApplicationConfigItemCommand'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ModifyWebApplicationConfigItemCommand'
      responses:
        '200':
          description: OK
  '/api/resource/WebApplicationResource/{id}/configuration/{configurationKey}':
    delete:
      tags:
        - WebApplicationResource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: configurationKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  '/api/resource/WebApplicationResource/{serverId}/containerAction':
    post:
      tags:
        - WebApplicationResource
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: actionId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
components:
  schemas:
    AuditLogEntry:
      required:
        - actionDisplayText
        - actionName
      type: object
      properties:
        id:
          type: string
          format: uuid
        actionName:
          type: string
        actionDisplayText:
          type: string
        actionMetadata:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        resourceId:
          type: string
          format: uuid
        resource:
          $ref: '#/components/schemas/BaseResource'
      additionalProperties: false
    AuditLogEntryPaginatedResponse:
      required:
        - data
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
      additionalProperties: false
    BaseResource:
      required:
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        auditLogEntries:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
      additionalProperties: false
    BaseResourceDto:
      required:
        - id
        - name
        - resourceType
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resourceType:
          type: string
      additionalProperties: false
    BaseResourceDtoPaginatedResponse:
      required:
        - data
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/BaseResourceDto'
      additionalProperties: false
...
    ContainerDto:
      type: object
      properties:
        stateRunning:
          type: boolean
        statePaused:
          type: boolean
        stateRestarting:
          type: boolean
        stateError:
          type: string
        stateStartedAt:
          type: string
          format: date-time
        stateFinishedAt:
          type: string
          format: date-time
      additionalProperties: false
...
    CreateWebApplicationResourceCommand:
      type: object
      properties:
        webApplicationName:
          type: string
        resourceGroupId:
          type: string
          format: uuid
        sourcePath:
          type: string
        sourceType:
          $ref: '#/components/schemas/WebApplicationSourceType'
      additionalProperties: false
...
    ModifyWebApplicationConfigItemCommand:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      additionalProperties: false
...
    PredefinedLokiQuery:
      enum:
        - ContainerLog
      type: string
    PredefinedPrometheusQuery:
      enum:
        - HttpRequestsCount
        - PostgresProcessesCount
        - PostgresEntriesInserted
        - PostgresEntriesReturned
        - GeneralCPULoad
        - GeneralMemoryUsage
      type: string
...
    PrometheusDataDto:
      type: object
      properties:
        resultType:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/PrometheusResultItemDto'
      additionalProperties: false
    PrometheusResultDto:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/PrometheusDataDto'
      additionalProperties: false
    PrometheusResultItemDto:
      type: object
      properties:
        metric:
          type: object
          additionalProperties:
            type: string
        stream:
          type: object
          additionalProperties:
            type: string
        values:
          type: array
          items:
            type: array
            items: { }
      additionalProperties: false
    PublicProxyConfigurationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        useHttps:
          type: boolean
        domain:
          type: string
        port:
          type: integer
          format: int32
      additionalProperties: false
    QueryLokiQuery:
      type: object
      properties:
        query:
          $ref: '#/components/schemas/PredefinedLokiQuery'
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        step:
          type: string
          nullable: true
        timeout:
          type: string
          nullable: true
        limit:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    QueryPrometheusQuery:
      type: object
      properties:
        query:
          $ref: '#/components/schemas/PredefinedPrometheusQuery'
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        step:
          type: string
          nullable: true
        timeout:
          type: string
          nullable: true
        limit:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
...
    UpdateWebApplicationDeploymentConfigurationCommand:
      type: object
      properties:
        buildCommand:
          type: string
        startupCommand:
          type: string
        runtimeType:
          $ref: '#/components/schemas/WebApplicationRuntimeType'
        port:
          type: integer
          format: int32
      additionalProperties: false
...
    VirtualNetworkSimpleDto:
      required:
        - id
        - name
        - resourceType
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resourceType:
          type: string
      additionalProperties: false
...
    WebApplicationConfigurationEntryDto:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      additionalProperties: false
    WebApplicationResourceDto:
      required:
        - id
        - name
        - resourceType
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resourceType:
          type: string
        resourceGroupId:
          type: string
          format: uuid
        resourceGroupName:
          type: string
        virtualNetworks:
          type: array
          items:
            $ref: '#/components/schemas/VirtualNetworkSimpleDto'
        publicProxyConfigurations:
          type: array
          items:
            $ref: '#/components/schemas/PublicProxyConfigurationDto'
        sourcePath:
          type: string
        sourceType:
          $ref: '#/components/schemas/WebApplicationSourceType'
        buildCommand:
          type: string
        startupCommand:
          type: string
        runtimeType:
          $ref: '#/components/schemas/WebApplicationRuntimeType'
        healthCheckUrl:
          type: string
        port:
          type: integer
          format: int32
        configuration:
          type: array
          items:
            $ref: '#/components/schemas/WebApplicationConfigurationEntryDto'
          nullable: true
      additionalProperties: false
    WebApplicationRuntimeType:
      enum:
        - Python
        - NodeJs
        - DotNet
      type: string
    WebApplicationSourceType:
      enum:
        - PublicGitClone
      type: string
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization token using the Bearer scheme. <br>\nEnter 'Bearer' [space] and then your token in the text input below. <br>\nExample: 'Bearer 12345abcdef'"
      name: Authorization
      in: header
security:
  - Bearer: [ ]